// frontend/src/api/model/responseBlogPostDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Rennan API
 * The Rennan API
 * OpenAPI spec version: 1.0.0
 */
import type { ResponseAccountDTO } from './responseAccountDTO'

export interface ResponseBlogPostDTO {
  /** Title of the BlogPost */
  title: string
  /** Content of the BlogPost */
  content: string
  /** Author id of the BlogPost */
  author_id: string
  /** Excerpt of the BlogPost */
  excerpt?: string
  /** Cover image of the BlogPost */
  cover_image?: string
  /** Published at of the BlogPost */
  published_at?: string
  /** Read time of the BlogPost */
  read_time?: number
  /** BlogPost id */
  id: number
  /** Author of the Post */
  author?: ResponseAccountDTO
  /** Date of creation */
  created_at: string
  /** Date of last update */
  updated_at: string
  /** Date of deletion */
  deleted_at: string
}