// frontend/src/api/model/responseBlogPostListDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Rennan API
 * The Rennan API
 * OpenAPI spec version: 1.0.0
 */
import type { ResponseBlogPostDTO } from './responseBlogPostDTO'
import type { ResponseMetaDTO } from './responseMetaDTO'

export interface ResponseBlogPostListDTO {
  /** List of BlogPost */
  data: ResponseBlogPostDTO[]
  /** Meta information */
  meta: ResponseMetaDTO
}