// frontend/src/api/model/responseCommentListDTO.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * Rennan API
 * The Rennan API
 * OpenAPI spec version: 1.0.0
 */
import type { ResponseCommentDTO } from './responseCommentDTO'
import type { ResponseMetaDTO } from './responseMetaDTO'

export interface ResponseCommentListDTO {
  /** List of Comment */
  data: ResponseCommentDTO[]
  /** Meta information */
  meta: ResponseMetaDTO
}